Rename vert-nav to horizontal-nav
Fixes https://jira.coreos.com/browse/CONSOLE-775
