+++ frontend/public/components/utils/_status-box.scss
@@ -34,7 +34,7 @@
   margin-bottom: 30px;
   text-align: center;
   .co-cluster-updates__component &,
-  .co-m-vert-nav & {
+  .co-m-horizontal-nav & {
     margin-top: 30px;
   }
 }
