+++ frontend/public/style.scss
@@ -72,7 +72,7 @@
 @import "components/toggle-play";
 @import "components/users";
 @import "components/volume-icon";
-@import "components/vert-nav";
+@import "components/horizontal-nav";
 @import "components/edit-yaml";
 @import "components/factory/modal-wrapper";
 @import "components/cluster-picker";
